AWS code restructured
