+++ record_audio.go
@@ -3,10 +3,6 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/bwmarrin/discordgo"
 	"github.com/joho/godotenv"
 	"github.com/pion/rtp"
@@ -19,10 +15,13 @@ import (
 	"syscall"
 )
 
-var ctx = context.Background()
-var dgVoice *discordgo.Session
-var connection *discordgo.VoiceConnection
-var fileLocation uint32
+var (
+	ctx = context.Background()
+	dgVoice *discordgo.Session
+	connection *discordgo.VoiceConnection
+	fileLocation uint32
+	)
+
 
 
 
@@ -175,54 +174,3 @@ func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
 
 
 
-//AWS - S3
-
-func ConnectAws() *session.Session {
-	AccessKeyID := GetEnvWithKey("AWS_ACCESS_KEY_ID")
-	SecretAccessKey := GetEnvWithKey("AWS_SECRET_ACCESS_KEY")
-	MyRegion := GetEnvWithKey("MyRegion")
-	sess, err := session.NewSession(
-		&aws.Config{
-			Region: aws.String(MyRegion),
-			Credentials: credentials.NewStaticCredentials(
-				AccessKeyID,
-				SecretAccessKey,
-				"",
-			),
-		})
-	if err != nil {
-		log.Println("Error :",err)
-	}
-	return sess
-}
-
-
-func AddFileToS3(s *session.Session, fileDir string) error {
-
-	file, err := os.Open(fileDir)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-	log.Println(GetEnvWithKey("S3Bucket"))
-	uploader := s3manager.NewUploader(s)
-	_, err = uploader.Upload(&s3manager.UploadInput{
-		Bucket:               aws.String(GetEnvWithKey("S3Bucket")),
-		Key:                  aws.String(fileDir),
-		Body:                 file,
-		ServerSideEncryption: aws.String("AES256"),
-	})
-	return err
-}
-
-func AddtoS3(fileDir string) error{
-	log.Println("FileDIR: ",fileDir)
-	s := ConnectAws()
-	// Upload
-	err := AddFileToS3(s, fileDir)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-	return nil
-}
